Fix bug when moving up and down from/to a "tabbed" line
Example:

test line 1
   ^--------- cursor here (E.cx = 3, E.rx = 3)
\t  test line 2

When we move down a line:

test line 1
\t  test line 2
      ^--------- cursor here (E.cx = 3, E.rx = 6)

Cursor shifts to the right. Because E.cx stays the same, but due
to the tab character, E.rx is calculated to be 6.

In this commit, after moving to a line, we update E.cx based on E.rx
(in text editors we move in screen space, not "character" space),
so in scroll(), E.rx is correctly updated from the new E.cx.

We can't directly update E.rx because it is thrown and computed every
refresh.